Rework sparse arrays.
Break sparse bit array class into two, making sparse array semi-abstract superclass. Introduct new sparse array subclass, sparse sample array. Add BASampleArray protocol, to which BASampleArray and BASparseSampleArray classes conform.

+++ BAFoundation/BASampleArray.m
@@ -66,7 +66,7 @@
 //        base[i] = sample[i];
 }
 
-static inline NSUInteger indexForCoordinates(NSUInteger *coordinates, NSUInteger power) {
+static inline NSUInteger indexForCoordinates(uint32_t *coordinates, NSUInteger power) {
     
     NSUInteger sampleIndex = 0;
     NSUInteger factor = 1;
@@ -78,11 +78,11 @@ static inline NSUInteger indexForCoordinates(NSUInteger *coordinates, NSUInteger
     return sampleIndex;
 }
 
-- (void)sample:(UInt8 *)sample atCoordinates:(NSUInteger *)coordinates {
+- (void)sample:(UInt8 *)sample atCoordinates:(uint32_t *)coordinates {
      [self sample:sample atIndex:indexForCoordinates(coordinates, _power)];
 }
 
-- (void)setSample:(UInt8 *)sample atCoordinates:(NSUInteger *)coordinates {
+- (void)setSample:(UInt8 *)sample atCoordinates:(uint32_t *)coordinates {
     [self setSample:sample atIndex:indexForCoordinates(coordinates, _power)];
 }
